Move Profile into a separate file
